+++ chocolatey-visualstudio.extension/CHANGELOG.md
@@ -1,13 +1,15 @@
 ﻿# CHANGELOG
 
 ## Version 1.5.0
-- New helper: Get-VisualStudioInstaller.
-  Retrieves an object containing the executable path and version number of the Visual Studio Installer (VS 2017+) installed on the system, if present.
 - New package parameter: '--layout D:\Path' can be used to create an offline installation source ("layout").
   Package installation using this parameter will throw an error at the end (code 814) so that Chocolatey does not register the package as installed.
   Supported by the Install-VisualStudio helper, both for Visual Studio 2017 and 2015.
 - New package parameter: '--bootstrapperPath D:\Path\vs_Enterprise.exe' can be used to install Visual Studio from a previously created offline installation source ("layout").
   Supported by the Install-VisualStudio helper, both for Visual Studio 2017 and 2015.
+- New helper: Get-VisualStudioInstaller.
+  Retrieves an object containing the executable path and version number of the Visual Studio Installer (VS 2017+) installed on the system, if present.
+- New helper: Install-VisualStudioInstaller.
+  Installs or updates the Visual Studio Installer (VS 2017+).
 
 ## Version 1.4.1
 - Fixed encoding of recently added files.